requirements.txt
crewai
crewai-tools
crewai[tools]
langchain
langchain-huggingface
load_dotenv
streamlit
